requirements.txt
Flask==2.3.2
azure-identity==1.13.0
azure-storage-blob==12.14.1
requests==2.28.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
